requirements.txt
pyspark
pandas
numpy
scikit-learn
mlflow
dvc[s3]
fastapi
uvicorn
pydantic
matplotlib
requests
PyYAML
